refactor: refactor LogfilesPanel.vue and split in multiple components
