+++ src/store/variables.ts
@@ -132,7 +132,7 @@ export const availableKlipperConfigReferenceTranslations = ['it', 'hu', 'zh']
 /*
  * List of all downloadable logfiles
  */
-export const genericLogfiles = ['crowsnest', 'mmu', 'sonar']
+export const genericLogfiles = ['klippy', 'moonraker', 'crowsnest', 'mmu', 'sonar']
 
 /*
  * List of all rollover logfiles
